Move tests to directory
